Move dokka to publishing plugin
