+++ build-logic/src/main/kotlin/com/gabrielfeo/published-kotlin-jvm-library.gradle.kts
@@ -1,9 +1,50 @@
+package com.gabrielfeo
+
+import org.jetbrains.dokka.DokkaConfiguration.Visibility.PUBLIC
+import org.jetbrains.dokka.gradle.DokkaTask
+import java.net.URL
+
 plugins {
     id("com.gabrielfeo.kotlin-jvm-library")
     `java-library`
     `maven-publish`
     signing
     id("org.jetbrains.kotlinx.binary-compatibility-validator")
+    id("org.jetbrains.dokka")
+}
+
+java {
+    withSourcesJar()
+    withJavadocJar()
+}
+
+val kotlinSourceRoot = file("src/main/kotlin")
+tasks.withType<DokkaTask>().configureEach {
+    dokkaSourceSets.all {
+        sourceRoot(kotlinSourceRoot)
+        sourceLink {
+            localDirectory = kotlinSourceRoot
+            remoteUrl = providers.gradleProperty("repo.url")
+                .map { URL("$it/blob/$version/${kotlinSourceRoot.relativeTo(rootDir)}") }
+            remoteLineSuffix = "#L"
+        }
+        jdkVersion = java.toolchain.version
+        suppressGeneratedFiles = false
+        documentedVisibilities = setOf(PUBLIC)
+        perPackageOption {
+            matchingRegex = """.*\.internal.*"""
+            suppress = true
+        }
+        externalDocumentationLink("https://kotlinlang.org/api/kotlinx.coroutines/")
+        externalDocumentationLink("https://square.github.io/okhttp/4.x/okhttp/")
+        externalDocumentationLink("https://square.github.io/retrofit/2.x/retrofit/")
+        externalDocumentationLink("https://square.github.io/moshi/1.x/moshi/")
+        externalDocumentationLink("https://square.github.io/moshi/1.x/moshi-kotlin/")
+    }
+}
+
+tasks.named<Jar>("javadocJar") {
+    from(tasks.dokkaHtml)
 }
 
 publishing {
